move aggregation stuff out of dashboard.py and add test
